requirements.txt
# Sentinel & Geospatial
sentinelhub
rasterio
s2cloudless
folium
pyproj
opencv-python-headless
google-cloud-storage

# ML & Preprocessing
numpy
pandas
scikit-learn==1.2.1
lightgbm
catboost
xgboost

# API / App
streamlit
requests
python-dotenv

# Visualization
matplotlib

# MLOps
mlflow==2.2.2
psycopg2-binary

# API
fastapi         # API framework
pytz            # time zone management
uvicorn         # web server

# tests
httpx           # HTTP client
pytest-asyncio  # asynchronous I/O support for pytest
